+++ arbiter-core/src/middleware.rs
@@ -31,7 +31,6 @@ use crate::environment::{EventBroadcaster, ResultReceiver, ResultSender, TxSende
 use crate::{
     agent::{Agent, NotAttached},
     environment::Environment,
-    utils::{recast_address, revm_logs_to_ethers_logs},
 };
 
 pub struct Connection {
@@ -256,8 +255,8 @@ impl Middleware for RevmMiddleware {
         let revm_result = self.provider().as_ref().result_receiver.recv().unwrap();
         let output = match revm_result.result.clone() {
             ExecutionResult::Success { output, .. } => output,
-            ExecutionResult::Revert { output, .. } => panic!("Failed due to revert: {:?}", output),
-            ExecutionResult::Halt { reason, .. } => panic!("Failed due to halt: {:?}", reason),
+            ExecutionResult::Revert { output, .. } => panic!("Failed due to revert: {:?}", output), // TODO: We can throw an error here instead and pause the environment if need be.
+            ExecutionResult::Halt { reason, .. } => panic!("Failed due to halt: {:?}", reason), // TODO: We can throw an error here instead and pause the environment if need be.
         };
         match output {
             Output::Create(bytes, ..) => {
@@ -320,4 +319,55 @@ impl Middleware for RevmMiddleware {
         let id = self.new_filter(FilterKind::Logs(filter)).await?;
         Ok(FilterWatcher::new(id, self.provider()).interval(Duration::ZERO))
     }
+}
+
+
+/// Recast a logs from Revm into the ethers.rs Log type.
+/// # Arguments
+/// * `revm_logs` - Logs from Revm. (Vec<revm::primitives::Log>)
+/// # Returns
+/// * `Vec<ethers::core::types::Log>` - Logs recasted into ethers.rs Log type.
+pub fn revm_logs_to_ethers_logs(
+    revm_logs: Vec<revm::primitives::Log>,
+) -> Vec<ethers::core::types::Log> {
+    let mut logs: Vec<ethers::core::types::Log> = vec![];
+    for revm_log in revm_logs {
+        let topics = revm_log.topics.into_iter().map(recast_b256).collect();
+        let log = ethers::core::types::Log {
+            address: recast_address(revm_log.address),
+            topics,
+            data: ethers::core::types::Bytes::from(revm_log.data),
+            block_hash: None,
+            block_number: None,
+            transaction_hash: None,
+            transaction_index: None,
+            log_index: None,
+            transaction_log_index: None,
+            log_type: None,
+            removed: None,
+        };
+        logs.push(log);
+    }
+    logs
+}
+
+// Certainly will go away with alloy-types
+/// Recast a B160 into an Address type
+/// # Arguments
+/// * `address` - B160 to recast. (B160)
+/// # Returns
+/// * `Address` - Recasted Address.
+pub fn recast_address(address: B160) -> Address {
+    let temp: [u8; 20] = address.as_bytes().try_into().unwrap();
+    Address::from(temp)
+}
+
+/// Recast a B256 into an H256 type
+/// # Arguments
+/// * `input` - B256 to recast. (B256)  
+/// # Returns
+/// * `H256` - Recasted H256.
+pub fn recast_b256(input: revm::primitives::B256) -> ethers::types::H256 {
+    let temp: [u8; 32] = input.as_bytes().try_into().unwrap();
+    ethers::types::H256::from(temp)
 }
